feat(sdk): expose spawn_stream and collect_candidates
